Final Product sans Runnable
Everything is finalized. However, the JAR will not work because the
audio and images folders are not in the same package as the
ResourceLoader, which prevents it from properly getting them. AS SOON AS
THAT IS FIXED, it will be considered complete.

Also we need to comment the hell out of it.